backend/pyproject.toml
[tool.poetry]
name = "carddemo-api"
version = "0.1.0"
description = "CardDemo mainframe migration — FastAPI backend"
authors = ["TrieDatum Engineering <engineering@triedatum.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.7.0"}
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
httpx = "^0.27.0"
aiosqlite = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
black = "^24.4.2"
ruff = "^0.4.5"
mypy = "^1.10.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
